cleanup canvas_animate
